Move the logic of motion-aware in tdc-peeler.
