+++ src/spikeinterface/sortingcomponents/matching/method_list.py
@@ -4,7 +4,6 @@ from .naive import NaiveMatching
 from .tdc import TridesclousPeeler
 from .circus import CircusPeeler, CircusOMPSVDPeeler
 from .wobble import WobbleMatch
-from .tdc_drift import TridesclousDriftPeeler
 
 matching_methods = {
     "naive": NaiveMatching,
@@ -12,7 +11,6 @@ matching_methods = {
     "circus": CircusPeeler,
     "circus-omp-svd": CircusOMPSVDPeeler,
     "wobble": WobbleMatch,
-    "tdc-drift-peeler": TridesclousDriftPeeler,
 }
 
 try:
