+++ src/spikeinterface/sortingcomponents/matching/tdc.py
@@ -9,6 +9,8 @@ from spikeinterface.core import (
 from spikeinterface.sortingcomponents.peak_detection import DetectPeakLocallyExclusive, DetectPeakMatchedFiltering
 from .base import BaseTemplateMatching, _base_matching_dtype
 
+from spikeinterface.generation.drift_tools import DriftingTemplates
+
 
 try:
     import numba
@@ -33,20 +35,27 @@ class TridesclousPeeler(BaseTemplateMatching):
     Contrary tp circus_peeler or wobble, this template matching is working directly one the waveforms.
     There is no SVD decomposition
 
-
+    A new mode motion_aware=False/True has been added to use a Motion (and/or DriftingTemplates)
+    for the peeler, this avoid using a interpolated recording. In that case the recording do not move
+    but the template are moving for the template matching.
     """
 
     def __init__(
         self,
         recording,
+        templates=None,
         return_output=True,
         parents=None,
-        templates=None,
         peak_sign="neg",
         exclude_sweep_ms=0.5,
         peak_shift_ms=0.2,
         detect_threshold=5,
         noise_levels=None,
+        # motion zone
+        motion_aware=False,
+        motion=None,
+        drifting_templates=None,
+        motion_step_um = 2.0,
         use_fine_detector=True,
         # TODO optimize theses radius
         detection_radius_um=80.0,
@@ -59,7 +68,10 @@ class TridesclousPeeler(BaseTemplateMatching):
         amplitude_limits=(0.7, 1.4),
     ):
 
-        BaseTemplateMatching.__init__(self, recording, templates, return_output=True, parents=None)
+        BaseTemplateMatching.__init__(self, recording, templates, return_output=return_output, parents=parents)
+
+
+        self.motion_aware = motion_aware
 
         unit_ids = templates.unit_ids
         channel_ids = recording.channel_ids
@@ -84,8 +96,10 @@ class TridesclousPeeler(BaseTemplateMatching):
         if s1 == 0:
             s1 = None
 
+        self.slice_short = slice(s0, s1)
+
         # TODO check with out copy
-        self.sparse_templates_array_short = templates.templates_array[:, slice(s0, s1), :].copy()
+        # self.sparse_templates_array_short = templates.templates_array[:, slice(s0, s1), :].copy()
 
         self.peak_shift = int(peak_shift_ms / 1000 * sr)
 
@@ -101,12 +115,71 @@ class TridesclousPeeler(BaseTemplateMatching):
         else:
             self.sparsity_mask = np.ones((unit_ids.size, channel_ids.size), dtype=bool)
 
+        self.motion = motion
+        if self.motion_aware:
+            if self.motion is None:
+                raise ValueError("TDC peeler : when using motion_aware=True, the motion must be given")
+
+            if drifting_templates is None:
+                # drifting template can be done externally to fasten the startup
+                self.drifting_templates = DriftingTemplates.from_static_templates(templates)
+                
+                min_, max_ = self.motion.get_boundaries()
+                steps = np.arange(min_, max_+motion_step_um/2, motion_step_um)
+                displacements = np.zeros((steps.size, 2), dtype="float64")
+                displacements[:, self.motion.dim] = steps
+                interpolation_kwargs = dict(interpolation_method="cubic")
+                self.drifting_templates.precompute_displacements(displacements, **interpolation_kwargs)
+            else:
+                self.drifting_templates = drifting_templates
+
+            # this is dense with shape (num_displacements, num_units, num_samples, num_channels)
+            templates_array_moved = self.drifting_templates.templates_array_moved
+            if templates.sparsity is not None:
+                # TODO later : move this logic into DriftingTemplate directly
+                max_num_active_channels = max(np.sum(self.sparsity_mask, axis=1))
+                sparsified_shape = templates_array_moved.shape[:-1] + (max_num_active_channels, )
+                self.sparse_templates_array_moved = np.zeros(shape=sparsified_shape, dtype=templates_array_moved.dtype)
+                for unit_index in range(unit_ids.size):
+                    chans = np.flatnonzero(self.sparsity_mask[unit_index, :])
+                    for d in range(templates_array_moved.shape[0]):
+                        sparsified = templates_array_moved[d, unit_index, :, :][:, chans]
+                        self.sparse_templates_array_moved[d, unit_index, :, :chans.size] = sparsified
+            else:
+                self.sparse_templates_array_moved = templates_array_moved
+
+            self.sparse_templates_array_static = None
+
+            # interpolation bins edges
+            interpolation_time_bin_size_s = 1.
+            self.interpolation_time_bins_s = []
+            self.interpolation_time_bin_edges_s = []
+            for segment_index, parent_segment in enumerate(recording._recording_segments):
+                # in this case, interpolation_time_bin_size_s is set.
+                s_end = parent_segment.get_num_samples()
+                t_start, t_end = parent_segment.sample_index_to_time(np.array([0, s_end]))
+                halfbin = interpolation_time_bin_size_s / 2.0
+                segment_interpolation_time_bins_s = np.arange(t_start + halfbin, t_end, interpolation_time_bin_size_s)
+                segment_interpolation_time_bin_edges_s = np.arange(
+                    t_start, t_end + halfbin, interpolation_time_bin_size_s
+                )
+                self.interpolation_time_bins_s.append(segment_interpolation_time_bins_s)
+                self.interpolation_time_bin_edges_s.append(segment_interpolation_time_bin_edges_s)
+
+        else:
+            self.sparse_templates_array_moved = None
+            self.interpolation_time_bins_s = None
+            self.interpolation_time_bin_edges_s = None
+            self.sparse_templates_array_static = templates.templates_array
+
+
         extremum_chan = get_template_extremum_channel(templates, peak_sign=peak_sign, outputs="index")
         # as numpy vector
         self.extremum_channel = np.array([extremum_chan[unit_id] for unit_id in unit_ids], dtype="int64")
 
         channel_locations = templates.probe.contact_positions
         unit_locations = channel_locations[self.extremum_channel]
+        self.channel_locations = channel_locations
 
         # distance between units
         import scipy
@@ -120,12 +193,23 @@ class TridesclousPeeler(BaseTemplateMatching):
             self.possible_clusters_by_channel.append(cluster_inds)
 
         # precompute template norms ons sparse channels
-        self.template_norms = np.zeros(num_templates, dtype="float32")
-        for i in range(unit_ids.size):
-            chan_mask = self.sparsity_mask[i, :]
-            n = np.sum(chan_mask)
-            template = templates.templates_array[i, :, :n]
-            self.template_norms[i] = np.sum(template**2)
+        if self.motion_aware:
+            num_displacements = self.drifting_templates.displacements.shape[0]
+            self.template_norms_moved = np.zeros((num_displacements, num_templates), dtype="float32")
+            for d in range(num_displacements):
+                for i in range(unit_ids.size):
+                    chan_mask = self.sparsity_mask[i, :]
+                    n = np.sum(chan_mask)
+                    template = self.sparse_templates_array_moved[d, i, :, :n]
+                    self.template_norms_moved[d, i] = np.sum(template**2)
+
+        else:
+            self.template_norms_static = np.zeros(num_templates, dtype="float32")
+            for i in range(unit_ids.size):
+                chan_mask = self.sparsity_mask[i, :]
+                n = np.sum(chan_mask)
+                template = self.sparse_templates_array_static[i, :, :n]
+                self.template_norms_static[i] = np.sum(template**2)
 
         #
         distances = scipy.spatial.distance.cdist(channel_locations, channel_locations, metric="euclidean")
@@ -182,6 +266,9 @@ class TridesclousPeeler(BaseTemplateMatching):
         self.peeler_margin = max(self.nbefore, self.nafter) * 2
         self.margin = max(self.peeler_margin, self.detector_margin0, self.detector_margin1)
 
+
+
+
     def get_trace_margin(self):
         return self.margin
 
@@ -190,30 +277,131 @@ class TridesclousPeeler(BaseTemplateMatching):
         # TODO check if this is usefull
         residuals = traces.copy()
 
+
+        if self.motion_aware:
+            # we need to split [start_frame, end_frame] into sub bins to match the motion
+
+            # see also interpolate_motion_on_traces() maybe factorize this trick
+            times = self.recording.sample_index_to_time(np.arange(start_frame, end_frame), segment_index=segment_index)
+            # print(traces.shape, times.shape, start_frame, end_frame, self.margin)
+            assert times.shape[0] == (traces.shape[0] - 2 *self.margin)
+
+
+            time_bin_edge = self.interpolation_time_bin_edges_s[segment_index]
+            interpolation_bin_inds = np.searchsorted(time_bin_edge, times, side="right") - 1
+            # the time bins may not cover the whole set of times in the recording,
+            # so we need to clip these indices to the valid range
+            n_bins = time_bin_edge.shape[0] - 1
+            np.clip(interpolation_bin_inds, 0, n_bins - 1, out=interpolation_bin_inds)
+
+
+            
+            
+            total_num_chans = self.channel_locations.shape[0]
+            interp_times = np.empty(total_num_chans)
+            interpolation_bins_here = np.arange(interpolation_bin_inds[0], interpolation_bin_inds[-1] + 1)
+            current_start_index = 0
+
+            # LOOP over interpolation bins
+            loop = []
+            for count, interp_bin_ind in enumerate(interpolation_bins_here):
+                # print("count", count, "interp_bin_ind", interp_bin_ind)
+
+                bin_time = self.interpolation_time_bins_s[segment_index][interp_bin_ind]
+                interp_times.fill(bin_time)
+                channel_motions = self.motion.get_displacement_at_time_and_depth(
+                    interp_times,
+                    self.channel_locations[:, self.motion.dim],
+                    segment_index=segment_index,
+                )
+
+                # if not self.motion_aware:
+                #     # TODO REMOVE this hack
+                #     channel_motions[:] = 0
+
+
+                # quick search logic to find frames corresponding to this interpolation bin in the recording
+                # quickly find the end of this bin, which is also the start of the next
+                next_start_index = current_start_index + np.searchsorted(
+                    interpolation_bin_inds[current_start_index:], interp_bin_ind + 1, side="left"
+                )
+                # frames_in_bin = slice(current_start_index, next_start_index)
+                # times vector is WITHOUT margin so need a shift in the slice
+                local_residuals = residuals[current_start_index:next_start_index+2*self.margin]
+
+                loop.append( (current_start_index, next_start_index+2*self.margin, channel_motions) )
+
+                current_start_index = next_start_index
+
+        else:
+            start = 0
+            stop = residuals.shape[0]
+            channel_motions = None
+            loop = [(start, stop, channel_motions)]
+
+
+        # # LOOP over interpolation bins
+        # for count, interp_bin_ind in enumerate(interpolation_bins_here):
+        #     # print("count", count, "interp_bin_ind", interp_bin_ind)
+
+        #     bin_time = self.interpolation_time_bins_s[segment_index][interp_bin_ind]
+        #     interp_times.fill(bin_time)
+        #     channel_motions = self.motion.get_displacement_at_time_and_depth(
+        #         interp_times,
+        #         self.channel_locations[:, self.motion.dim],
+        #         segment_index=segment_index,
+        #     )
+
+        #     if not self.motion_aware:
+        #         # TODO REMOVE this hack
+        #         channel_motions[:] = 0
+
+
+        #     # quick search logic to find frames corresponding to this interpolation bin in the recording
+        #     # quickly find the end of this bin, which is also the start of the next
+        #     next_start_index = current_start_index + np.searchsorted(
+        #         interpolation_bin_inds[current_start_index:], interp_bin_ind + 1, side="left"
+        #     )
+        #     # frames_in_bin = slice(current_start_index, next_start_index)
+        #     # times vector is WITHOUT margin so need a shift in the slice
+        #     local_residuals = residuals[current_start_index:next_start_index+2*self.margin]
+
         all_spikes = []
-        level = 0
-        spikes_prev_loop = np.zeros(0, dtype=_base_matching_dtype)
-        use_fine_detector_level = False
-        while True:
-            # print('level', level)
-            spikes = self._find_spikes_one_level(residuals, spikes_prev_loop, use_fine_detector_level, level)
-            if spikes.size > 0:
-                all_spikes.append(spikes)
+        for start, stop, channel_motions in loop:
 
-            level += 1
+            local_residuals = residuals[start:stop]
 
-            # TODO concatenate all spikes for this instead of prev loop
-            spikes_prev_loop = spikes
+            spikes_in_time_bin = []
+            level = 0
+            spikes_prev_loop = np.zeros(0, dtype=_base_matching_dtype)
+            use_fine_detector_level = False
+            while True:
+                # print('level', level)
+                spikes = self._find_spikes_one_level(local_residuals, spikes_prev_loop, use_fine_detector_level, level, channel_motions)
+                if spikes.size > 0:
+                    spikes_in_time_bin.append(spikes)
 
-            if (spikes.size == 0) or (level == self.max_peeler_loop):
-                if self.use_fine_detector and not use_fine_detector_level:
-                    # extra loop with fine detector
-                    use_fine_detector_level = True
-                    level = self.max_peeler_loop - 1
-                    continue
-                else:
-                    break
+                level += 1
 
+                # TODO concatenate all spikes for this instead of prev loop
+                spikes_prev_loop = spikes
+
+                if (spikes.size == 0) or (level == self.max_peeler_loop):
+                    if self.use_fine_detector and not use_fine_detector_level:
+                        # extra loop with fine detector
+                        use_fine_detector_level = True
+                        level = self.max_peeler_loop - 1
+                        continue
+                    else:
+                        break
+            
+            for spikes in spikes_in_time_bin:
+                spikes["sample_index"] += start
+            
+            all_spikes.extend(spikes_in_time_bin)
+
+            
+            
         if len(all_spikes) > 0:
             all_spikes = np.concatenate(all_spikes)
             order = np.argsort(all_spikes["sample_index"])
@@ -223,7 +411,8 @@ class TridesclousPeeler(BaseTemplateMatching):
 
         return all_spikes
 
-    def _find_spikes_one_level(self, traces, spikes_prev_loop, use_fine_detector, level):
+    def _find_spikes_one_level(self, traces, spikes_prev_loop, use_fine_detector, level, channel_motions):
+        
 
         # print(use_fine_detector, level)
 
@@ -239,7 +428,9 @@ class TridesclousPeeler(BaseTemplateMatching):
         else:
             peak_detector = self.fast_spike_detector
 
+        # print('peak_detector', peak_detector)
         detector_margin = peak_detector.get_trace_margin()
+
         if self.peeler_margin > detector_margin:
             margin_shift = self.peeler_margin - detector_margin
             sl = slice(margin_shift, -margin_shift)
@@ -251,6 +442,9 @@ class TridesclousPeeler(BaseTemplateMatching):
         peak_sample_ind = peaks["sample_index"]
         peak_chan_ind = peaks["channel_index"]
         peak_sample_ind += margin_shift
+
+        # print()
+        # print('peaks', peaks.size)
 
         peak_amplitude = traces[peak_sample_ind, peak_chan_ind]
         order = np.argsort(np.abs(peak_amplitude))[::-1]
@@ -278,16 +472,39 @@ class TridesclousPeeler(BaseTemplateMatching):
             sample_index = peak_sample_ind[i]
 
             chan_ind = peak_chan_ind[i]
-            possible_clusters = self.possible_clusters_by_channel[chan_ind]
 
+            if self.motion_aware:
+                local_motion = np.zeros(2, dtype=self.channel_locations.dtype)
+                local_motion[self.motion.dim] = channel_motions[chan_ind]
+
+                # move this channel to the original position
+                peak_location_moved = self.channel_locations[chan_ind, :] - local_motion
+                # print(self.channel_locations)
+                chan_ind_moved = np.argmin(np.sum((self.channel_locations - peak_location_moved)**2, axis=1))
+                # if np.sum(local_motion) != 0:
+                #     print("local_motion", local_motion, "chan_ind", chan_ind, "chan_ind_moved", chan_ind_moved)
+                displacement_index = np.argmin(np.sum((self.drifting_templates.displacements - local_motion)**2, axis=1))
+                templates_array = self.sparse_templates_array_moved[displacement_index, :, :, :]
+
+                template_norms = self.template_norms_moved[displacement_index, :]
+            else:
+                chan_ind_moved = chan_ind
+                templates_array = self.sparse_templates_array_static
+
+                template_norms = self.template_norms_static
+
+
+            possible_clusters = self.possible_clusters_by_channel[chan_ind_moved]
+
+            # shorten in time
+            sparse_templates_array_short = templates_array[:, self.slice_short, :]
 
             if possible_clusters.size > 0:
                 cluster_index = get_most_probable_cluster(
                     traces,
-                    self.sparse_templates_array_short,
+                    sparse_templates_array_short,
                     possible_clusters,
                     sample_index,
-                    chan_ind,
                     self.nbefore_short,
                     self.nafter_short,
                     self.sparsity_mask,
@@ -298,7 +515,8 @@ class TridesclousPeeler(BaseTemplateMatching):
                 # chans = np.any(self.sparsity_mask[possible_clusters, :], axis=0)
                 # wf = traces[sample_index - self.nbefore : sample_index + self.nafter][:, chans]
                 # ax.plot(wf.T.flatten(), color='k')
-                # dense_templates_array = self.templates.get_dense_templates()
+                # # dense_templates_array = self.templates.get_dense_templates()
+                # dense_templates_array = self.drifting_templates.templates_array_moved[displacement_index,:, :, :]
                 # for c_ind in possible_clusters:
                 #     template = dense_templates_array[c_ind, :, :][:, chans]
                 #     ax.plot(template.T.flatten())
@@ -312,7 +530,7 @@ class TridesclousPeeler(BaseTemplateMatching):
                 # find best shift
                 numba_best_shift_sparse(
                     traces,
-                    self.sparse_templates_array_short[cluster_index, :, :],
+                    sparse_templates_array_short[cluster_index, :, :],
                     sample_index,
                     self.nbefore_short,
                     self.possible_shifts,
@@ -338,17 +556,16 @@ class TridesclousPeeler(BaseTemplateMatching):
                         spikes[i]["cluster_index"] == spikes_prev_loop[b]["cluster_index"]
                     ):
                         is_valid = False
-
+                # print(is_valid)
                 if is_valid:
                     # temporary assign a cluster to neighbors if not done yet
                     inner_neighbors_inds = [ind for ind in neighbors_spikes_inds[i] if (ind > i and ind < spikes.size)]
                     for b in inner_neighbors_inds:
                         spikes["cluster_index"][b] = get_most_probable_cluster(
                             traces,
-                            self.sparse_templates_array_short,
+                            sparse_templates_array_short,
                             possible_clusters,
                             spikes["sample_index"][b],
-                            spikes["channel_index"][b],
                             self.nbefore_short,
                             self.nafter_short,
                             self.sparsity_mask,
@@ -359,8 +576,8 @@ class TridesclousPeeler(BaseTemplateMatching):
                         spikes[inner_neighbors_inds],
                         traces,
                         self.sparsity_mask,
-                        self.templates.templates_array,
-                        self.template_norms,
+                        templates_array,
+                        template_norms,
                         self.nbefore,
                         self.nafter,
                     )
@@ -372,7 +589,7 @@ class TridesclousPeeler(BaseTemplateMatching):
                         construct_prediction_sparse(
                             spikes[i : i + 1],
                             traces,
-                            self.templates.templates_array,
+                            templates_array,
                             self.sparsity_mask,
                             wanted_channel_mask,
                             self.nbefore,
@@ -464,7 +681,6 @@ def get_most_probable_cluster(
     sparse_templates_array,
     possible_clusters,
     sample_index,
-    chan_ind,
     nbefore_short,
     nafter_short,
     template_sparsity_mask,
